+++ libcaf_core/caf/all.hpp
@@ -39,6 +39,7 @@
 #include "caf/config_value_reader.hpp"
 #include "caf/config_value_writer.hpp"
 #include "caf/const_typed_message_view.hpp"
+#include "caf/cow_tuple.hpp"
 #include "caf/deep_to_string.hpp"
 #include "caf/defaults.hpp"
 #include "caf/deserializer.hpp"